add some build/destroy scripts

+++ destroy.sh
@@ -14,8 +14,22 @@ PREFIX=$(awk -F\" '/^prefix/{print $2}'  "environments/$CONFIG.tfvars")
 AWS_DCOS_STACK_BUCKET=$(awk -F\" '/^dcos_stack_bucket/{print $2}'  "environments/$CONFIG.tfvars")
 AWS_DCOS_APPS_BUCKET=$(awk -F\" '/^dcos_apps_bucket/{print $2}'  "environments/$CONFIG.tfvars")
 
-aws s3 rm "s3://$AWS_DCOS_STACK_BUCKET" --recursive
-aws s3 rm "s3://$AWS_DCOS_APPS_BUCKET" --recursive
+if aws s3 ls "s3://$AWS_DCOS_STACK_BUCKET" 2>&1 | grep -q 'NoSuchBucket'
+  then
+    echo "No bucket to delete"
+else
+  echo "Deleting bucket $AWS_DCOS_STACK_BUCKET"
+  aws s3 rm "s3://$AWS_DCOS_STACK_BUCKET" --recursive
+fi
+
+if aws s3 ls "s3://$AWS_DCOS_APPS_BUCKET" 2>&1 | grep -q 'NoSuchBucket'
+  then
+    echo "No bucket to delete"
+else
+  echo "Deleting bucket $AWS_DCOS_STACK_BUCKET"
+  aws s3 rm "s3://$AWS_DCOS_APPS_BUCKET" --recursive
+fi
+
 
 STACKS=("${PREFIX}platform" "${PREFIX}redshift" "${PREFIX}vpc" "${PREFIX}iam")
 for i in "${STACKS[@]}"; do
